[extractor] Add magentamusik.de
