+++ yt_dlp/extractor/_extractors.py
@@ -1013,7 +1013,7 @@ from .lynda import (
 )
 from .m6 import M6IE
 from .magellantv import MagellanTVIE
-from .magentamusik360 import MagentaMusik360IE
+from .magentamusik import MagentaMusikIE
 from .mailru import (
     MailRuIE,
     MailRuMusicIE,
